refactor: rename Canvas.tsx to CursorField.tsx
Disambiguates the low-level cursor renderer from panel-level components
that share "Canvas" in their name (ReactionCanvas, ImageCanvas,
ArrivalCanvas, etc.). All imports, stories, and tests updated.
