+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file. Releases cu
 - **Map server tests updated for `viewerConfig`** — `createState` and `getPersistedState` assertions now include the `viewerConfig` field added in the last release cycle.
 
 ### Changed
+- **`Canvas.tsx` renamed to `CursorField.tsx`** — disambiguates the low-level cursor renderer from the many panel-level components that use "Canvas" in their names (`ReactionCanvas`, `ImageCanvas`, `ArrivalCanvas`, etc.). All imports, stories, and tests updated.
 - **`activity` state replaced by `screenPanels` map; `ReactionCanvasParticipant` no longer always-mounted** — `ReactionCanvasAppV4` replaces the `activity: string` state with `screenPanels: Record<string, string>` (currently `{ canvas: panelId }`), making the single canvas screen explicit and extensible. Activity changes (`activityChanged` server message and initial `currentActivity` on connect) are now handled via a `useMessageSubscription` in the app shell rather than bubbling up through `Canvas` → `ReactionCanvasParticipant` callbacks, so the canvas component no longer needs to stay mounted when another interface is active. `ReactionCanvasParticipant` is now conditionally rendered only when the canvas interface is visible. `ActivityMode` type alias retired throughout app, party server, and plugins; all usages replaced with `string`.
 - **Typecheck and tests gate PR staging deploys** — `staging-deploy.yml` now runs a `check` job (typecheck + Storybook/vitest tests) before deploying to staging; deploy is blocked if checks fail.
 - **All panel plugins migrated to shared `RoomSocketContext`** — 13 components (`VoiceCallPanel`, `LightShow`, `ScreenLight`, `MapMaker`, `MapViewer`, `Steno`, `StoryTracer`, `NeighborPanel`, `SignatureCanvas`, `HelloWorld`, `MoodTones`, `ArrivalCanvas`, plus `ReactionCanvasParticipant`) now consume the shared participant socket from `RoomSocketProvider` lifted to `ReactionCanvasAppV4`, eliminating per-panel socket connections and their reconnect latency on activity switch. `MoodTones` and `ArrivalCanvas` previously connected as admin unnecessarily; both now use the participant socket. Storybook `preview` gains a global `RoomSocketProvider` decorator so plugin stories render without individual wrappers.
