+++ app/components/apps/ReactionCanvasAppV5.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from "react";
-import Canvas from "../shared/Canvas";
+import Canvas from "../shared/CursorField";
 import TouchLayer from "../shared/TouchLayer";
 import { RoomSocketProvider } from "../../contexts/RoomSocketContext";
 import AdminPanelWithDB from "../panels/AdminPanelWithDB";
