+++ tests/CursorField.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, act } from '@testing-library/react'
-import Canvas from '../app/components/shared/Canvas'
+import CursorField from '../app/components/shared/CursorField'
 import { RoomSocketProvider } from '../app/contexts/RoomSocketContext'
 
 const mockSend = vi.hoisted(() => vi.fn())
@@ -34,24 +34,24 @@ beforeEach(() => {
   capturedOnMessage.fn = null
 })
 
-describe('Canvas socket behaviour', () => {
+describe('CursorField socket behaviour', () => {
   it('calls onSocketReady with a send function on mount', () => {
     const onSocketReady = vi.fn()
-    renderWithProvider(<Canvas userId="user1" onSocketReady={onSocketReady} />)
+    renderWithProvider(<CursorField userId="user1" onSocketReady={onSocketReady} />)
     expect(onSocketReady).toHaveBeenCalledOnce()
     expect(typeof onSocketReady.mock.calls[0][0]).toBe('function')
   })
 
   it('calls onPresenceCount when presenceCount message is received', () => {
     const onPresenceCount = vi.fn()
-    renderWithProvider(<Canvas userId="user1" onPresenceCount={onPresenceCount} />)
+    renderWithProvider(<CursorField userId="user1" onPresenceCount={onPresenceCount} />)
     act(() => emitMessage({ type: 'presenceCount', count: 42, viewerCount: 3 }))
     expect(onPresenceCount).toHaveBeenCalledWith(42)
   })
 
   it('calls onConnected with inviteEdges and currentActivity from connected message', () => {
     const onConnected = vi.fn()
-    renderWithProvider(<Canvas userId="user1" onConnected={onConnected} />)
+    renderWithProvider(<CursorField userId="user1" onConnected={onConnected} />)
     act(() => emitMessage({
       type: 'connected',
       inviteEdges: { user2: 'user1' },
@@ -61,7 +61,7 @@ describe('Canvas socket behaviour', () => {
   })
 
   it('opens socket as admin when readOnly is true', () => {
-    renderWithProvider(<Canvas userId="user1" />, { readOnly: true })
+    renderWithProvider(<CursorField userId="user1" />, { readOnly: true })
     expect(capturedConfig.current.query).toEqual({ isAdmin: 'true' })
   })
 })
